Add Android example "helloWorld" app
This compiles and runs, but is obviously not ready for pants yet. I intend
to use it as the platform to develop the base required tasks.

In the future, I will add more to it as I need to express more and more
tasks. Some updated styling will eventually be included for pride reasons,
I pretty much wrote this of the top of my head and it looks like Android 4.1.1 ;)

Pants copyright headers are included on all files, I believe.