Add android_resources target
The syntax of this addition to the BUILD file is still WIP.
The aapt tool spiders downwards and wants a top level directory as an
argument. But pants uses the leaves for BUILD files, as does javac.

The location of the output R.Java will need to be passed as the injected
dep into Pants' dependency graph, and the path is also an argument to javac.
So the best way to declare and parse the resources is still under
consideration.