Add "home" field to android targets.
This, like the "package" field, can be parsed from available information
and so may be excised from final targets. But unlike the "package"
field, this would be pretty ugly to parse. I am considering adding a
BUILD file for apks, which would live at the project root. But that
seems counter to the pants BUILD file idea, so for right now, it is
a manual field.