First gathering of review files
This contains the integration example app, and targets for the first
stage of android targets.

The android_disribution simply passes on any SDK it sees declared in
the users path. The android_resources target is acted upon by the
aapt_gen task. This spiders down the res folder and converts the xml
and res into an R.java class.

Ideally, this should be able to be ran on the example app included.
ATM there is some installation errors.