Correctly add requirements of included dependencies that provide distributions.
Testing Done:
When we upgraded from twitter.pants to pantsbuild.pants, setup_py broke for us in that not all source distributions had the correct set of install_requires dependencies.  This elif clause was in the original setup_py so I'm not sure why it was removed.

Submitted on behalf of Brian Wickman
Reviewed at https://rbcommons.com/s/twitter/r/566/