Add android sample app and install pants.backend.android into pants
This is not complete. This only is working when PANTS_DEV=1. Without it,
pants cannot find a module android.register. My best guess after working
on it for quite awhile is that there is a difference in the PYTHONPATH
in DEV mode that allows it to find the module.

It is basically working in DEV mode, but I need to figure out how to
install the module in the PYTHONPATH.