Add JarLibrary to create_target deps
The JarLibrary has a jar field filled by a
JarDependency object. I was trying to make an actual
JarDependency target, but everything was
failing further and further backwards.

Still not compiling. I need the android.jar on the
classpath, I fear that this is not what the deps do.
I am plumbing JavaLibrary to see what happens.

Cleanup from adding JarLibrary target to deps

This looks to be about right, but now to plump JavaLibrary
and see why the deps aren't on the classpath.