Added RepeatingAlarmSample from AOSP
This is an AOSP sample app that usually comes with a gradle build
setup. I am using it to test our Android support against maven-style
layouts, and to do CI on how hard it is to port a gradle build to
a pantsbuild. So far, so good.

This still won't compile for a couple reasons. The java_library
deps need access to android.jars. Now this could possibly be
sorted by adding JarDep just like I do in aapt_gen. But better
to have an SDK resolver that is able to resolve known jars from the
Android SDK. This would takes care of the second issue, which is
that this sample app requires a standard android_library that comes
with the SDK, the support-lib. These deps can be fetched by Maven,
but we will need to map the source included in the SDK to our
current AndroidDistribution object.