Add a default definition to pants.ini
We read the keystore names and config file in the SignApk class.
This is done so we can keep the reading of pants.ini restricted
to Tasks code. I assign the results to the targets themselves
so that I can pass the target as the argument to the KeyResolver.
I could be persuaded that it would be better to pass
(config_file, keystore) tuples to KeyResolver instead. But as it
stands having the instantiated Keystore objects associate with
the specific target will make actual signing pretty easy.