Reading the config file is workin
As is, it is reading the definition somewhat like we would want.
But it is reading in the dictionary keys as strings, when it would
be much better to have them as variables. But since they will not
be defined as variables in the config file, they need to be strings
for now. Pretty ugly in several probably unnecessary ways.