Working with simplified definition
About to muck with it again though. I want the list of
keystore names to be optional. They are too crazy looking
to make as a requirement each build.

Ideally, the user can decide to pass a list of specific keystore
names they would like to sign with. But if they just pass a config
file name, Pants should be smart enough to assume that they would
like to sign the package with each key defined in the config file.

So I am reorganizing again to work toward that.