Moving keystore_config location to the options framework
As per some feedback on the design doc. This is moving all keystore
defs away from the BUILD and into the options system. That allows
config files to be defined in pants.ini, pantsrc, CLI, or env. variables.

Also will be moving from pants.Config to python's ConfigParser on Eric's
advice.