TestKeystoreResolver class is moving along
The Config.get_required is not robust against empty string fields,
as in a section is defined but with no value. If the get_required
call expects a string, it interprets the absence of a value as just
an empty string. The question is whether or not I want to allow that
behavior (keystores with a blank alias field, perhaps) or push a patch
and get consensus.

dont know yet.