Change the KeyResolver to return a list of keys instead of a dict
The dict was cumbersome and unneeded since we no longer are using
a list of names to verify which keys we want.

If we ever go back to explicitly naming per-project keys, then perhaps
I will have to go back to the dict. For now, this is cleaner and more
intuitive.