Change Distribution.cached to compare using Revision objects
When spinning up distributions, Distribution.cached returns
previously instantiated jdk distros. The proper distro to return
was determined by comparing min/max arguments. The problem is that
cached() is generally passed strings . Those strings were being
compared with Revision objects. This should have raised an
exception b/c Revision.__cmp__ is overloaded so as to
compare _components.

This exception was never raised though, because Pants was almost
always just returning the Distribution that was cached during
setup. The _CACHE.key was just (minimum_version, jdk) which
is (None, True) or (None, False) for all situations I could find.

This conspired to cause pants to not respect any max_version
constraints. The solution is to cast the max/min version args
of Distribution.cached to full Revision objects for easy
apples-to-apples comparison. The _CACHE.key has been expanded to
(minimum_version, maximum_version, jdk) so as to honor all
constraints.