always use relpath on --test file args, refactor execute in junit_run, add more integration tests
Specifying the test to run prefixed with ./ fails because the ./ causes the src path to not match the paths in the classes_by_source product. This changes that so it will always create the relpath when trying to derive the classes.

It also refactors execute to reduce nesting and clarify some of the flag / config interactions.

Testing Done:
wrote regression test that fails, changed implementation to make it pass. CI baking here https://travis-ci.org/pantsbuild/pants/builds/55514326

Bugs closed: 1313

Reviewed at https://rbcommons.com/s/twitter/r/1976/