Add android_library support and android ivysettings
Android has 3 different concepts of android_library:
.jars, .aars and source files. This first step only
expressly supports from jars. Support for source
libraries should be a fairly easy followup. .aars are TODO.

This includes several separate example projects that
will probably not need to be merged into Pants itself.
The best thing would be for me to write a couple new
Hello-Pants sample apps strictly for Pants, but time
is short. The are all Apache or MIT licensed code and
the LICENSES are included, so whatever works.

There are two new local ivy resolvers. One is to interact
with the local Android SDK as set by env ANDROID_HOME.
The second is for jars I have checked-in to 3rdParty.
The 3rdParty checking jars should be able to go away
whenever someone gets a chance to add .aar support.

The ivysettings for the local Android SDK is currently
unused, because fetching jars violates the new classpath
rules. I am including it here because it probably is
the way we will want to interact with the SDK so I
wanted to hear some thioughts as to a good solution.

The Android BUILD file format needs a reimagining
after library support goes in as well. Some of the
things the BUILD file requires could be easily
reduced to sensible defaults. A TODO as well.

Compiling any of the included examples worka but
you have to pass compile.checkstyle --skip for a couple
of them.