It is working, believe it or not.
The fix was in the dependency injection.
When using codegen, it was never injecting the JavaLibrary
dep because the AndroidBinary (the gentarget) didn't have any
dependees. I think that moving off codegen is the best move.
The codegen class doesn't have any first-rate mechanism to create
targets for each created R.Java. This allows us to call create_target
for each. The issue that leaves is how to best hook into the
invalidation framework.

To be continued, but now with 100% less despair.