Substantially cleaned version of the working code
Needs to be optimized some, I am idly hoping I can reduce
the calls to target.walk to just one. We will see.

But everything is working from what I can tell. The big
issue is that if same library is compiled against multiple
sdks, then it produces multiple R.java files, as it should.
But that chokes the javac compile, because it records multiple
versions of the same class. So that needs to be thought about
(or most likely asked about, I am already over time here).