Adding the aapt tests.
Everything is gren except for the render args stuff. It could
be easily mad to pass, but really it needs to be reworked.

The big things to test are the things that I have been
checking as I have been writing...are the deps handled
correctly and do the proper commands get executed.