POC for unpacked aar
This class.jar is just from the root of the unpacked .aar. I checked
it into libs manually, but it compiles and runs.

So ivy fetches the .aar, I need to hook into the resole phase
and unpack the jar.

I also need to find out if I can use the R.txt (I bet I include the
res in the aaptBuild and the R.txt in the compile.)