POC for unpacking .aar
This goes pretty much as expected, actually. The last mystery
I left in place, so I don't forget about it. To integrate the
google_play_services android library you usually
would import the entire google-play-services_lib, which has
a resource for the google_play_services_version. I have
not found where to get that with the .aar.

But if I just delete the requirement from the app's manifest
everything compiles and runs. Too bad that isn't a final solution.