Beginning of adding an explode_aar task
We actually cannot do this in the resolve phase as originally
planned because some of the exploded arr contents are bound
as codegen targets. This tentatively installs the explode_aars
task in the import phase and models it after the unpack_jars task.

Upcoming is the new targets. Some of this has been on the menu for
awhile, so hopefully this will make some sense. I think I have it
whittled to just two new targets. We'll see.