Unpacking aars manually with the intercepted includes_pattern
This has the android_library intercepting the includes_pattern
kwarg before calling the unpack_jars super class. This is done
so as to treat the android_deps artifacts the same, whether
jar or aar.

This still requires a ton of thought. But it is working for now.
I may end up pulling the android_library off of unpacked_jars
completely. We'll see.