Stashing some progress before I go crazy
There is a certain level of gymnastics needed to have the
android_dep work for both jars and aars. Especially since
they actually need to be treated pretty differently.

If I want to have the includes/excludes work I need to be
thoughtful. An 'include' on a jar means that the classes.jar
of the aar never gets unpacked. But preemptively grabbing
the includes before it gets to unpacked means that there
is no working includes on the jar.

So, conundrum. I have some ideas, hence the stash.