Move all unpacked jar contents under target.id
This is safe because each spec has a unique include/exclude
pattern. So everything that depends on a particular spec
will be able to use the exact same unpacked dir.

This has the negative effect of possibly unpacking
a library multiple times, if one library is defined
by multiple android_library targets.

A solution to the latter problem could be moving the filter
process to dx compile. By filtering the extracted output
we would increase the load of dx but decrease the load
on explode_aars (soon to be renamed).