Fixed GMS and it works
Pulling the google-play-services aar pulls in all the deps in the
pom. They are all unpacked properly.

This alerted me to the possibility that some aars will not have
a classes.jar (contrary to google's spec). So all the classes.jar
stuff got moved under the if block.

The code rightly iterates over every artifact pulled in by the
library. It then needs to make a Jar target for each artifact.
That is next.