Everything is basically working, except for the manifest issue.
Essentialy, manifests are not required for aar files, they are
contained within the binary. So I need to make manifests optional
for certain android targets.

This will require creating a flag for certain targets to raise or
not raise exceptions if the manifest is not declared. For aar
files, the manifest will be set in explode_aar task.

The first consumer of manifests is in aapt_gen, so an exception
can be put their to raise if any target has not had its manifest
resolved by that point.