Okay, this works but...
I have been adding JarLibrary and AndroidResource targets.
That is incorrect. I need to be adding an AndroidLibrary target
for every aar. Essentially I am converting each aar to a jar
library. Doing this in the explode_aars file just obfuscates
that from the user base.

Of course I am sensitive to the argument that they should be
separate targets. As of now they are not, lets see how crazy
this has to be before I feel compelled to change.