Beginning of moving the filter function to dx task
This has some benifits, but perhaps some hidden costs. In a better
world I would try it both ways and profile.

The advantages are: only one new target per android_dependency. Each
android_dependency gets unpacked only once, which is the biggest.
If a repo had 20 targets that depended on google-play-services
(the pom of which has ~13 aar files) each aar would be unpacked
for each target. This is due to the includes/excludes, I need
to be safe for different targets to have different includes/excludes.

If we unpack the entire library then it becomes a common unit. The
point where we need them to differentiate from each other is in the
dx step, where we pack the dex file with the classes. So I have
moved the filtering to that point in the build.

As of now I am importing the class methods in unpack_jars. Those
functions could maybe get moved up to fs or util. It doesn't make
the best sense where it is as of now.