Apps build and run now, with products passed to dxcompile.
There is so much hackery, though. The fundamental problem
solved by this commit is creating a product from explode and
then retrieving it in DxCompile. The compliations are that
a jar could actually be pulled in by several AndroidDependency
definitions. A good example is support-v4 which might be
declared as a separate dep but is also pulled in by the google-
play-services pom. So that jar is depended upon several
different places in the graph. The challenge is to not unpack
it multiple times(handled) and also to not add it to the
Dx invocation more than once (which causes an error).

The hacky solution here is to have a set of relative filenames
(relative in that they match the include/exclude patterns) and
membership in added classes is checked before adding to the
classes.jar set.

Lots of work to do, lots of cleanup. But after I add the
create_targets stuff to Explode, the major framework will be up.