Libraries are working.
Lots of things to check. You cannot compile all android
targets because aapt_gen tries to add the same android.jar to
the graph multiple times. Which shouldn't be the case.

ALso, I think that the classpath will need to be isolated so
that if android_targets depend on different versions of the
android.jar then it can be passed to the compiler separately.

With the batch job, pants complains that the class has already
been added.