With a bunch of hacks - everything is building
However --

Two big issues remain: As of right now, android_dependencies cannot
be added as simple deps for AndroidBinaries. You must create an
android_library container and set the android_dependency as its
library. Some thought is needed to decide how best to handle that
or perhaps to just leave that as is for now.

The second is bigger: for some reason the applications that use
google-play-services are crashing saying they cannot instantaite
a class. I see that the proper class is being sent to dx and is
included in the apk. My hope is that it is merely being
put in the wrong spot. I am going to build it with another
working branch and compare the decompiled output.