Okay, everything is working so cleanup/refactor is starting
There is still a runtime crash in apps that use the aar libraries.
I am mostly unsure of what is causing that. But it could be related
to needing an aidl task or handling the other jars that aars pack in
the libs folder.

The included example apps run so I am going to make that runtime
crash a TODO. This is a HUGE patch as is, and I need to get it in
since I am waaaaay past my personal deadline.