Move ExplodeAar to ExplodeLibraries and install in unpack-jars phase
The explode class is no longer a subclass of UnpackJars since
it doesn't filter the output. It only uses the unpacking methods
from fs. I think many of the methods in unpack jars should be moved
to fs.archive, I added that as a TODO and will see what the
feedback is like for that idea.