Add invalidation to explode_libraries and conitnue cleanup
The invalidation looks to be working fine for this class. I
experimented with pulling the ImportMixin out AndroidLibrary.
But I am unsure how to do that at this moment. I use imports
to get the jar files to operate on them. Otherwise I am stuck
with just target graph nodes.

Maybe someone in feedback will have a simple method.