Working on deduping archives per-binary
This change is to be smarter about deduping dependencies.
THe Dx tool exits with an error if multiple copies of the same file
are in the class files to be packed into classes.dex. Originally,
this change just upated a set to ensure that property. But that
doesn't account for circumstances where the classes may be
different versions of each other.

Now that we use the .m2 repo, it is very easy to unknownigly
pull in dependencies. For instance, pulling in the gms library
now also pulls in the support-v4 library by default. Importing
version 22 of the gms library also pulls in version 22 of the suport
library. The trouble comes if perhaps you already have a explicit
dependency on version 21 of the support library. Under the old
system this is a silent error. Which version of the support
classes end up in the dex is a matter of circumstance.

The plan is to gather all the archives per binary, and dedup them.
It will probably be common for binaries to have mutliple copies
of common libraries. Pants already dedupes the actual artifacts
what we are doing is handling this for the dex step.

Since all copies of the same jar are now removed, if a class is added
to the dex file a second time, that means that the apk is relying
on two different versions of the same library. In that case
we want to raise an exception and advise them to be more explicit with
their android_dependencies.