Stash some rough drafts.
I think I am going to go back to the old way. I cannot
arbitrarily dedupe per archive because which include/exclude
pattern should I use in this case? I think it would
be best to just raise a noisy exception in Dx if the class
is already added. I was trying to be flexible with non-specific
maven fetching, but I think that just bring pain.

The future changes are - instead of using just updating the set()
of class files, I will raise a DxError if a duplicated class is
noticed.

Providing a specific error message will be a pain, and I may have
to punt on that issue for now.