Deduping checks are working
This was actually a pretty big problem. Interacting with the
builtin m2 repos of the SDK, it is easy to pull in lots of
deps automatically. It would be easy to have multiple
versions of a dependency, e.g. an explicit dep on support-v4
version 21 as well as a dep on google-play-services 7.0.0. The gms
automatically brings in support-v4, so then you would have support-v4
21 and 22 as deps. So we want to allow duplicate deps if the version
number is the same, but raise an exception in the ambiguous case
of having deps on different versions.

This check is done in the lowest impact way I could find. We check
every added class ('org/pantsbuild/package/class.class). Only if
we detect that that class has already been added, then we check against
the full path of the duplicated class. If the full path matches,
 then we know that the version number is the same (due to the way
we handle unpacking libraries). If the path does not match, then
we have a version conflict and need to raise the Exception.