Commented and doctringed.
It needs to be linted --- and tested. I hope to have some sort of
working test structure in the next couple days. My dream would
be to have this in on Tuesday or wednesday.

We will see how that plan meshes with the fact that I am currently
on vacation with my family...