Add an android_target contextmanager to android_test_base
As advised by jsirois, this reduces the boilerplate. Looks like
adding a package_name kwarg to the manifest was easy too, we will
see if it works in practice. I don't have a specific use case
in mind but since we parse that manifest for targe attributes,
it stands to reason that we may want to vary those valus for tests.