Some progress on test_unpack_libraries
Several of the create_target methods are green. I need to add
tests for create_android_library and a smoke test.

The harder tests will be in unpack_integration. That is where
I will create some mock products and test to see what new targets
are actually placed in context.

I also need to add tests to DxCompile where I demonstrate that
the filtering is working. There should probably be added tests
to DxCompileIntegration, but I don't know if I will get to that.