Tests for unpack_libraries coming along
I changed the create_android_library_target to accept the location
of the unpacked library. It didn't need that information but testing
is way easier this way. Probably a little bit of the easy way out.

The unit tests are mostly done but only because I have decided to
push most of the heavy testing into the integration tests. The
integration test is running, now I need to test the output and
prove the invalidation is working.