Added compressed aarfile and unpacked aars.
This looks like it will work to pass as artifacts to android_dependency.
Not that that will be easy...not quite sure what I will need to do.

But this is a fairly flexible framework that should allow me to test
various scenarios w/o too much trouble.

A good followup to this patch would be a refactor of unpack_jars,
unpack_libraries, test_unpack_jars and test_unpack_libraries.
There is a bunch of code in the unpack_jars files that I had to
modify just enough to where I could not resuse of subclass it.
It would be good to further abstract some of the methods and
pull them up into fs or base for wider reuse.

In the future, maybe.