Now using the ivy_task_mixin mapjar names for archives in tests
Ivy transforms the jar name into a unique per-jar filename that
includes the version number. The UnpackLibraries task relies on
those two properties. Therefore the renaming behavior had to be
approximated in the test class.

As a regression test, I pulled the ivy_args in IvyMixinTask into
a static function and test for changes in the UnpackLibraries
unit tests. This is just to signal interest in the ivy_args,
the details are unimportant as long as the aforementioned
properties are maintained.