Add a second sample_aarfile
When working on this I found a bug in the original unpack_libraries
task. It keeps track of the unpacked libraries only per instance
of the class. So multiple executes in a row will perhaps unpack
the same target multiple times.

The example case for multiple unpacks is adding a new target that
depends on an already unpacked jar. The invalidation works only
on targets and the task just caches unpacked_archives per instance.
So in that case the archive is unneseccarily unpacked.

I am going to try and revamp the unpack_libraries. Let's see if I
can get the unpack half on the execute to work on AndroidDeps and
be per archive. The second half can continue to be on a
per AndroidLibrary basis.