Some more DxCompile cleanup
I think I am going to test the filter and exceptions in
integration tests. This is mostly driven by the fact that
creating dummy rpducts for classes and unpacked_archives
is such a pain. I could do it for unpacked_libraries but
it looks like a dummy classes_by_target would be a real
pain.