Grabbing the UnpackJarsFingerprintStrategy for UnpackLibraries
The two unpack classes _really_ need a refactor but it doesn't seem
fair to drag a whole bunch of working code into this massive Android
change.

Anyway, using the fingerprint strategy fixes the invalidation bugs.
Now if an AndroidDependency version is bumped, the target is back in
the active queue. All libraries are still only unpacked once per
archive, although if a new target is added that uses an existing
AndroidDependency, I think it will be unpacked a second time (or
will it? I may write a test for that but time is short)

Also, this had a crazy side effect where all the test_unpack_libraries
tests are green. About to investigate but I wanted to commit this
for posterity.