Import foursquare's buildgen code from Fsq.io as a contrib module
Foursquare is OSS-ing the buildgen code. This is a test
run over the pants base. Here we import the code and
and do a little fiddling with some contrib plugin BUILD defs.

I would like to publish buildgen as contrib modules, but it
looks like those are really only publishable from a pants
checkout. So we will see what distribution ends up looking
like. For now, we are going to try and run buildgen as
a loose source plugin imported directly into the
pants tree.