Results of running Python buildgen over Pants codebase.
Foursquare is opensourcing our Buildgen implementation.
The buildgen module was refactored to be per-language,
and I imported it into a contrib module and ran it over pants.

We hope to release the source this week - but in the meantime
a good test of the backend after the OSS-ing refactor was to
run it over the pants codebase.

The invocation was:

  ./pants buildgen src:: tests:: testsprojects:: examples:: \
          pants-plugins:: 3rdparty:: contrib::

Four concessions did I make to get the incantation to run:

* The old contrib/buildgen was not included - I deleted that backend
   to make the importing of our buildgen manageable.
* Temporarily delete the contrib/go backend.
   * It is undeleted in this commit - but the buildgens
     were clashing.
* Disable 'test_antlr_builder' and 'test_thrift_namespace_packages'
   * These can be made to work with buildgen by creating
     buildgen_target_bags - bags that aggregate targets for things
     like codegen.
   * But since they are both marked 'xxx' I decided to see if people
     thought that was worth the time.
* Move twitter.commons from 3rdparty/python/twitter/common into plain
  3rdparty/python
  * 3rdparty deps are mapped in a third_party_map Foursquare maintains.
    We share this map, but don't have any BUILD files in that directory.
    If pants maintains its own map or inlines a copy of ours, then
    the map can be adjusted to any 3rdparty layout pants prefers.

The contrib modules and the disabled tests were returned to their
original state before committing.