Buildgen the plugin code.
Buildgen is working. The first step for upstream is
probably top land a patch that buildgens the current
codebase. We can decide if/when/how upstream wants
to consume buildgen at our leisure.

So this is just a buildgen of the buildgen plugin -
this code doesn't need to import pants modules
so these imports were not in the BUILD files.