Rework CacheManager for clarity and correctness.
Before this commit, the VT and VTS classes set and reset lots of
instance attributes in order to determine whether a side effect like
create_results_dir had happened.

This reduces (but does not eliminate) that mutation. The chief remaining
offender is self.force_invalidate, which basically just resets self.valid to False
and potentially costs us some guarantees depending on when it is called.

There are no changes to public API. I settled on using two wrappers around result_dir
types. The difference between $type_results_path and $type_results_dir is that *path
is a potential location, while the *dir has been verified as existing.

IMO, every property except for the VTS.results_dir could be private and then missing
dirs could bubble up. Ideally, it would have some first-class way of detecting between
disabled caching and an internally-derived exception.

The changes include:
* localizing the hashing function in the CacheKeyGenerator,
* Renaming "current_results_dir" to "unique_results_dir"
  * (using "current" was easily conflated with the 'current' name for the stable dir)
* passing the task.workdir when instatiating the cache_manager
* removing vts.is_incremental, since that is currently only set per-task

There are some correctness improvements that fell out of this.
* one of the test_task tests was incorrectly interpreting the previous_results_dir, although that looks
to have minimal crossover with behavior in production.
* unique_result_dirs are wiped for invalid targets, ensuring that the results of a canceled run
or a failed resolve are not accidentally included in the next run.
* Believed that calling safe_mkdir(results_dir, clean=True) will now be safe, although always a
  noop.
    * The directory will always exist if caching is enabled, but the clean will never run since the
      the results_dir will always be empty.