Fork the junit-runner to a new namespace so we don't poison the namespace
I couldn't figure out anything simple here. I would have preferred to add
a classifier or something but I couldn't find anything obvious.

The junit-runner dependencies are also not published at their current
SemVer, so the transitive deps fail when publishing the runner alone.

At the end of the day, I munged the namespace of everything in the
junit-runner's closure and uploaded to our internal jar server.
All of the *fs jars are only available internally.

The patch we are backporting is already on master but we may
have already released the 1.4.0 final, which means we are stuck
with the forked jars until we consume 1.5.0, which is the biggest
breaking update since the 1.2.0 upgrade.

So:

1. ./pants publish --publish-jar-local=~/.m3 --publish-jar-force --no-publish-jar-dryrun --publish-jar-override=org.pantsbuild#junit-runner-fs=1.0.24  --publish-jar-override=org.pantsbuild#junit-runner-withretry-fs=0.0.17 --publish-jar-override=org.pantsbuild#junit-runner-annotations-fs=0.0.21  --publish-jar-override=org.pantsbuild#args4j-fs=0.0.19 --no-publish-jar-local-snapshot  src/java/org/pantsbuild::

2. Upload the jars manually to Nexus.
- This might have been easier to port the publish xml from our
internal repo, but I didn't want to risk accidentally pushing to
the upstream Pants maven account.

The upload process:
  - Go to Nexus
  - Log in
  - Repositories tab
    - 3rdparty Releases
    - Upload Artifact tab in lower pane
        - Pom
           - GAV Definition: From Pom
           - Select pom from disk
           - DONT UPLOAD YET
        - Jars
           - Select Jars to upload - 3x required:
               - jar, sources.jar, javadoc.jar
               - Just let autofill, ignore weirdness
           - Add Artifacts for all 3 in a row
       - NOW click upload

2. I also ported the foursquare/ivysettings.xml chain repos so that Pants
could test the resolution and run unit tests. If you messed up any
of the steps above you may see ClassNotFound and you should click
around the 3rdparty Releases repository.

Be aggressive with local cleans - you could really confuse yourself!