[Internal Port] Javadoc task should skip .java sources owned by scala_library
ScalaLibrary targets can safely consume mixed language source code,
a la `globs('*.scala', '*.java')`.

The JavadocGen task cannot operate on scala code, so it should either
tighten the targets it accepts with a target filter or otherwise
opt-out of operating on those targets.

It is more than a little weird for the Javadoc task to import a Scala target.
The better fix would be to use or create a first-class way of determining
java-only Java targets. A hackier way I considered was to use the `target_alias`
of `scala_library` but this was the easiest path that would not thrash during
updates.

https://github.com/foursquare/pants/commit/da179c8b7450d735d9d5a97ffa268690d745952f